models/egorov/requirements.txt
torch==1.10.0
absl-py==0.13.0
aioredis==1.3.1
alabaster==0.7.12
amply==0.1.4
amqp==5.0.6
asgiref==3.4.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
cachetools==4.2.2
catboost==0.26
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==1.6.0
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
daphne==3.0.2
deap==1.3.1
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.44.0
django-celery-beat==2.2.1
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-registration==3.2
django-resized==0.3.11
django-timezone-field==4.2.1
djangorestframework==3.12.4
docutils==0.17.1
et-xmlfile==1.1.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
graphviz==0.17
grpcio==1.39.0
h5py==3.1.0
hiredis==2.0.0
hpsklearn==0.1.0
hyperlink==21.0.0
hyperopt==0.2.5
idna==3.2
imageio==2.9.0
imagesize==1.2.0
incremental==21.3.0
Jinja2==3.0.1
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==5.1.0
lightgbm==3.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
msgpack==1.0.2
networkx==2.6.2
nltk==3.6.2
nose==1.3.7
numpy==1.19.5
numpydoc==1.1.0
oauthlib==3.1.1
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.1
patsy==0.5.1
Pillow==8.3.1
plotly==5.1.0
prompt-toolkit==3.0.19
protobuf==3.17.3
PuLP==2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pympler==1.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyschedule==0.2.34
python-crontab==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
randomcolor==0.4.4.5
redis==3.5.3
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
service-identity==21.1.0
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
statsmodels==0.12.2
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.2.0
tqdm==4.62.0
transliterate==1.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
xgboost==1.2.1
xlrd==2.0.1
XlsxWriter==1.4.5
zope.interface==5.4.0
psutil==5.8.0
django-filter~=2.4.0
django-multiselectfield==0.1.12
pm4py==2.2.15
kneed==0.7.0
kaggle==1.5.12
yfinance==0.1.64
Quandl==3.7.0
paramiko==2.7.2
olefile==0.46
TPOT==0.11.7
opencv-python==4.5.5.62
imutils==0.5.4
autokeras==1.0.15
auto-sklearn==0.14.7
geopandas==0.10.2
shapely==1.8.1.post1
flatten-json==0.1.13